add roles and respective functionalities

+++ app/Http/Controllers/CategoryController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Category;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Inertia\Inertia;
 
 class CategoryController extends Controller
@@ -13,6 +14,7 @@ class CategoryController extends Controller
      */
     public function index()
     {
+        Gate::authorize('verifyAdmin');
         return Inertia::render('Categories/Categories',[
             'categories'=>Category::all(),
         ]);
@@ -23,6 +25,10 @@ class CategoryController extends Controller
      */
     public function create()
     {
+        Gate::authorize('verifyAdmin');
+        return Inertia::render('Categories/NewCategory',[
+            'categories'=>Category::all(),
+        ]);
     }
 
     /**
@@ -30,6 +36,7 @@ class CategoryController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('verifyAdmin');
         $data = $request->all();
         Category::create($data);
         return to_route('categories.index');
@@ -48,6 +55,7 @@ class CategoryController extends Controller
      */
     public function edit(Category $category)
     {
+        Gate::authorize('verifyAdmin');
         return Inertia::render('Categories/EditCategory',[
             'categories'=>Category::find($category->id),
         ]);
@@ -58,6 +66,7 @@ class CategoryController extends Controller
      */
     public function update(Request $request, Category $category)
     {
+        Gate::authorize('verifyAdmin');
         Category::find($category->id)->update($request->all());
         return to_route('categories.index');
     }
@@ -67,6 +76,7 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
+        Gate::authorize('verifyAdmin');
         $category = Category::find(($category->id));
         $category->delete();
         return to_route('categories.index');
